Coded ActorFactory, still needs testing
Today, I added TinyXml-2 support in order to use it for my actor system.
so far, I have the headers for Actor, ActorFactory, and ActorComponent
coded up, and I made an initial attempt at my ActorFactory class. I'm
probably going to want to do quite a bit of testing on the ActorFactory
since there seem to be a lot of things that could go wrong with it.