Coded VisualComponent, basic sprite rendering
Made a lot of progress this commit. Most significantly, coded up the
VisualComponent class, cleaned up some namespace issues, and implemented
a basic sprite-rendering system. As of now, the inner-game state has a
sprite logo which has an impulse applied to it and thus, moves towards
the lower right corner of the screen. Now that I have a visual system
running, the next step is to test the collision system that I wrote,
using the visual sprites.